Refactor var Lexer.tokens & func is_keyword
